slight refactor and added a test to ensure the new, more generic code produces the same result as the previous version compute_int_samples_v0.py (the literal translation of Qingge's fortran code)
